# common-platform-product-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-product-mgmt
  description: Product Management Core Application
  contact:
    name: Catalis Development
    email: dev@catalis.com
  version: 1.0.0
servers:
  - url: 'http://core.catalis.vc/products'
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
  '/api/v1/products/{productId}/versions/{versionId}':
    get:
      tags:
        - Product Version
      summary: Get Product Version by ID
      description: >-
        Retrieve a specific product version using its unique identifier,
        ensuring it belongs to the specified product.
      operationId: getProductVersion
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: integer
            format: int64
        - name: versionId
          in: path
          description: Unique identifier of the product version
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved the product version record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductVersionDTO'
        '404':
          description: Product version not found
    put:
      tags:
        - Product Version
      summary: Update Product Version
      description: Update an existing version record associated with the specified product.
      operationId: updateProductVersion
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: integer
            format: int64
        - name: versionId
          in: path
          description: Unique identifier of the product version record to update
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductVersionDTO'
        required: true
      responses:
        '200':
          description: Product version updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductVersionDTO'
        '404':
          description: Product version not found
    delete:
      tags:
        - Product Version
      summary: Delete Product Version
      description: Remove an existing product version record by its unique identifier.
      operationId: deleteProductVersion
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: integer
            format: int64
        - name: versionId
          in: path
          description: Unique identifier of the product version to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Product version deleted successfully
        '404':
          description: Product version not found
...
  '/api/v1/products/{productId}':
    get:
      tags:
        - Product
      summary: Get Product by ID
      description: Retrieve a specific product by its unique identifier.
      operationId: getProduct
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved the product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '404':
          description: Product not found
    put:
      tags:
        - Product
      summary: Update Product
      description: Update the information of an existing product by its unique identifier.
      operationId: updateProduct
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product to update
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
        required: true
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '404':
          description: Product not found
    delete:
      tags:
        - Product
      summary: Delete Product
      description: Remove an existing product by its unique identifier.
      operationId: deleteProduct
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Product deleted successfully
        '404':
          description: Product not found
...
  '/api/v1/products/{productId}/versions/filter':
    post:
      tags:
        - Product Version
      summary: List Product Versions
      description: >-
        Retrieve a paginated list of all versions associated with the specified
        product.
      operationId: getAllProductVersions
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaginationRequest'
        required: true
      responses:
        '200':
          description: Successfully retrieved the product versions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseProductVersionDTO'
        '404':
          description: No product versions found for the specified product
  '/api/v1/products/{productId}/versions':
    post:
      tags:
        - Product Version
      summary: Create Product Version
      description: Create a new product version for a specific product.
      operationId: createProductVersion
      parameters:
        - name: productId
          in: path
          description: Unique identifier of the product
          required: true
          schema:
            type: integer
            format: int64
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductVersionDTO'
        required: true
      responses:
        '200':
          description: Product version created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductVersionDTO'
        '400':
          description: Invalid product version data provided
...
  /api/v1/products/filter:
    post:
      tags:
        - Product
      summary: List Products
      description: Retrieve a paginated list of products.
      operationId: getAllProducts
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: >-
            Unique key for idempotent requests. If provided, ensures that
            identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaginationRequest'
        required: true
      responses:
        '200':
          description: Successfully retrieved the list of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseProductDTO'
        '404':
          description: No products found
...
components:
  schemas:
    ProductVersionDTO:
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        productVersionId:
          type: integer
          format: int64
          readOnly: true
        productId:
          type: integer
          format: int64
        versionNumber:
          type: integer
          format: int64
        versionDescription:
          type: string
        effectiveDate:
          type: string
          format: date-time
...
    ProductDTO:
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        productId:
          type: integer
          format: int64
          readOnly: true
        productSubtypeId:
          type: integer
          format: int64
        productType:
          type: string
          enum:
            - FINANCIAL
            - NON_FINANCIAL
        productName:
          type: string
        productCode:
          type: string
        productDescription:
          type: string
        productStatus:
          type: string
          enum:
            - PROPOSED
            - ACTIVE
            - RETIRED
        launchDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: 'The direction of sorting, either ASC or DESC.'
          example: DESC
          default: DESC
      description: >-
        Represents a pagination request for retrieving paginated results,
        including page number, size, sort field, and direction.
    PaginationResponseProductVersionDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/ProductVersionDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: 'The current page number, typically zero-based.'
          format: int32
      description: >-
        Represents a paginated response containing a list of items and
        pagination metadata.
...
    PaginationResponseProductDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/ProductDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: 'The current page number, typically zero-based.'
          format: int32
      description: >-
        Represents a paginated response containing a list of items and
        pagination metadata.
...